v210enc,dnxhdenc,dvenc,rawvideoenc: set bitrate avcodec: print correct ratecontrol mode in dump_format avcodec: change bitrate related option min value to 0 avcodec: change AVCodecContext and AVFormatContext bit_rate field to int64_t needed for rawvideo

+++ libavcodec/mpegvideo_enc.c
@@ -364,7 +364,7 @@ av_cold int MPV_encode_init(AVCodecContext *avctx)
         av_log(avctx, AV_LOG_INFO, "impossible bitrate constraints, this will fail\n");
     }
 
-    if(avctx->rc_buffer_size && avctx->bit_rate*(int64_t)avctx->time_base.num > avctx->rc_buffer_size * (int64_t)avctx->time_base.den){
+    if(avctx->rc_buffer_size && avctx->bit_rate*avctx->time_base.num > avctx->rc_buffer_size * (int64_t)avctx->time_base.den){
         av_log(avctx, AV_LOG_ERROR, "VBV buffer too small for bitrate\n");
         return -1;
     }
